refactor: removed user and page authentication logic from auth.ts
